+++ cascade-oj/ent/judgerecord/where.go
@@ -11,57 +11,57 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.JudgeRecord {
+func ID(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.JudgeRecord {
+func IDEQ(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.JudgeRecord {
+func IDNEQ(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.JudgeRecord {
+func IDIn(ids ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.JudgeRecord {
+func IDNotIn(ids ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.JudgeRecord {
+func IDGT(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.JudgeRecord {
+func IDGTE(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.JudgeRecord {
+func IDLT(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.JudgeRecord {
+func IDLTE(id int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldLTE(FieldID, id))
 }
 
 // ProblemID applies equality check predicate on the "problem_id" field. It's identical to ProblemIDEQ.
-func ProblemID(v int) predicate.JudgeRecord {
+func ProblemID(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldProblemID, v))
 }
 
 // UserID applies equality check predicate on the "user_id" field. It's identical to UserIDEQ.
-func UserID(v int) predicate.JudgeRecord {
+func UserID(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldUserID, v))
 }
 
@@ -81,42 +81,42 @@ func Code(v string) predicate.JudgeRecord {
 }
 
 // ProblemIDEQ applies the EQ predicate on the "problem_id" field.
-func ProblemIDEQ(v int) predicate.JudgeRecord {
+func ProblemIDEQ(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldProblemID, v))
 }
 
 // ProblemIDNEQ applies the NEQ predicate on the "problem_id" field.
-func ProblemIDNEQ(v int) predicate.JudgeRecord {
+func ProblemIDNEQ(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNEQ(FieldProblemID, v))
 }
 
 // ProblemIDIn applies the In predicate on the "problem_id" field.
-func ProblemIDIn(vs ...int) predicate.JudgeRecord {
+func ProblemIDIn(vs ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldIn(FieldProblemID, vs...))
 }
 
 // ProblemIDNotIn applies the NotIn predicate on the "problem_id" field.
-func ProblemIDNotIn(vs ...int) predicate.JudgeRecord {
+func ProblemIDNotIn(vs ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNotIn(FieldProblemID, vs...))
 }
 
 // UserIDEQ applies the EQ predicate on the "user_id" field.
-func UserIDEQ(v int) predicate.JudgeRecord {
+func UserIDEQ(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldEQ(FieldUserID, v))
 }
 
 // UserIDNEQ applies the NEQ predicate on the "user_id" field.
-func UserIDNEQ(v int) predicate.JudgeRecord {
+func UserIDNEQ(v int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNEQ(FieldUserID, v))
 }
 
 // UserIDIn applies the In predicate on the "user_id" field.
-func UserIDIn(vs ...int) predicate.JudgeRecord {
+func UserIDIn(vs ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldIn(FieldUserID, vs...))
 }
 
 // UserIDNotIn applies the NotIn predicate on the "user_id" field.
-func UserIDNotIn(vs ...int) predicate.JudgeRecord {
+func UserIDNotIn(vs ...int64) predicate.JudgeRecord {
 	return predicate.JudgeRecord(sql.FieldNotIn(FieldUserID, vs...))
 }
